Restructure workspace and driver packages
