+++ test/integration/mount_nonexistent_path_test.go
@@ -18,13 +18,13 @@ import (
 	"testing"
 
 	"github.com/oracle/oci-flexvolume-driver/pkg/flexvolume"
-	"github.com/oracle/oci-flexvolume-driver/pkg/oci/driver"
+	"github.com/oracle/oci-flexvolume-driver/pkg/oci/driver/block"
 )
 
 // TestMountNonexistentPath checks an error is return when the required mout
 // point does not exist during mounting.
 func TestMountNonexistentPath(t *testing.T) {
-	d := &driver.OCIFlexvolumeDriver{}
+	d := &block.Driver{}
 	opts := flexvolume.Options{
 		"kubernetes.io/fsType":         "ext4",
 		"kubernetes.io/pvOrVolumeName": fw.VolumeName,
